Make docker2aci a library
